NCKUBOT/requirements.txt
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.2
future==0.18.2
gunicorn==20.0.0
idna==2.8
imgurpython==1.1.7
line-bot-sdk==1.8.0
psycopg2==2.8.4
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.7
